Cleaned up code deleted non needed files and added delete options for post and shopping products
