+++ src/app/core/services/likes/likes.service.ts
@@ -3,6 +3,7 @@ import {BehaviorSubject} from 'rxjs';
 import {AngularFirestore} from '@angular/fire/firestore';
 import {post} from 'selenium-webdriver/http';
 import {take} from 'rxjs/operators';
+import {Post} from '../../models/Post';
 
 @Injectable({
   providedIn: 'root'
@@ -50,9 +51,9 @@ export class LikesService {
     const postRef = this._afStore.doc(`Users/${userId}/post/${postId}`);
     this._afStore.firestore.runTransaction(t => {
       return t.get(postRef.ref).then(doc => {
-        console.log(doc.data().likeCount);
-        if (doc.data().likeCount) {
-          const newCount = doc.data().likeCount + 1;
+        const p = <Post> doc.data();
+        if (p.likeCount) {
+          const newCount = p.likeCount + 1;
           return t.update(postRef.ref, {
             likeCount: newCount
           });
@@ -70,9 +71,9 @@ export class LikesService {
     this._afStore.firestore.runTransaction(t => {
       return t.get(postRef.ref)
         .then(doc => {
-          console.log(doc.data().likeCount);
-          if (doc.data().likeCount) {
-            const newCount = doc.data().likeCount - 1;
+          const p = <Post> doc.data();
+          if (p.likeCount) {
+            const newCount = p.likeCount - 1;
             t.update(postRef.ref, {
               likeCount: newCount
             });
