--- src/app/pages/content/content.component.ts
@@ -1,137 +1,0 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
-import { Post } from './post';
-import { AngularFirestore, AngularFirestoreDocument, AngularFirestoreCollection } from '@angular/fire/firestore';
-import { CdkVirtualScrollViewport } from '@angular/cdk/scrolling';
-import { isFakeMousedownFromScreenReader } from '@angular/cdk/a11y';
-import * as faker from '../../../app/faker.js';
-import { ActivatedRoute } from '@angular/router';
-import { AuthService } from '../../core/services/auth/auth.service';
-import {MatDialog, MatDialogRef} from '@angular/material';
-import {PostDialogComponent} from './post-dialog/post-dialog.component';
-
-@Component({
-  selector: 'content',
-  templateUrl: './content.component.html',
-  styleUrls: ['./content.component.scss']
-})
-export class ContentComponent implements OnInit {
-  htmlele: HTMLElement;
-  posts: any;
-  batch = 20;
-  @ViewChild(CdkVirtualScrollViewport)
-  viewport: CdkVirtualScrollViewport;
-  faker = faker;
-  public uidParam;
-  private auth: any;
-  snapshot;
-  postDoc: AngularFirestoreDocument<Post>;
-  postDialogRef: MatDialogRef<PostDialogComponent>;
-  postsCollection: AngularFirestoreCollection<Post>;
-
-  newContent = "type note here";
-
-  constructor(
-    private db: AngularFirestore,
-    private _route: ActivatedRoute,
-    private _authService: AuthService,
-    private _dialog: MatDialog
-  ) {
-    this.uidParam = this._route.snapshot.params['id'];
-  }
-
-  ngOnInit() {
-    this.auth = this._authService.getAuth().value;
-    this.postsCollection = this.db.collection(`Users/${this.uidParam}/post`, ref => ref.orderBy('date', 'desc'));
-    this.posts = this.postsCollection.valueChanges();
-    // this.snapshot = this.postsCollection.snapshotChanges().subscribe(res => {
-    //   this.snapshot = res;
-    //   console.log(this.snapshot);
-    // });
-    // this.faker.personas();
-  }
-
-  subscribe() {
-    this.hide;
-  }
-
-  hide(id) {
-    this.htmlele = document.getElementById(id) as HTMLElement;
-    this.htmlele.parentNode.removeChild(this.htmlele);
-  }
-
-  trackByIdx(i) {
-    return i;
-  }
-
-  openDialog() {
-    this.postDialogRef = this._dialog.open(PostDialogComponent, {
-      data: {
-        auth: this.auth,
-      }
-    });
-  }
-
-  addPost() {
-    let randomName = faker.name.findName();
-    let paragraph = faker.lorem.paragraph();
-    let profileUrl: string = faker.random.number() % 5 === 0 ?  'https://i.imgur.com/y45bKaJl.jpg' : faker.internet.avatar();
-    profileUrl = 'https' + profileUrl.substring(5);
-    let postType = faker.random.number() % 2 === 0 ? 'text' : 'img';
-    console.log(profileUrl);
-    console.log(randomName);
-    console.log(paragraph);
-    let date = new Date();
-    let postToAdd: Post = new Post(randomName, profileUrl, date.toISOString(), paragraph, faker.random.image(), postType, '');
-    // "", this.newContent, , 'incomplete', 'yellow');
-
-
-    this.postsCollection.add({ ...postToAdd });
-    // this.notesCollection.get()
-  }
-
-  displayId(x) {
-    console.log(x);
-  }
-
-  changeColor(index, color) {
-    // get noteRef at this id
-    let id = this.snapshot[index].payload.doc.id;
-    this.postDoc = this.db.doc('posts/' + id);
-    // subscribe to this note, cast it as a new one, and update it
-    let newPost: Post;
-    this.postsCollection.doc(id).get().subscribe(res => {
-      newPost = res.data() as Post;
-      newPost.color = color;
-      this.postDoc.set(newPost);
-
-    });
-
-
-  }
-  deleteNote(index) {
-    // get noteRef at this id
-    //console.log(this.uidParam)
-    let id = this.snapshot[index].payload.doc.id;
-    this.postsCollection.doc(id).delete();
-  }
-
-  liked(index) {
-    let id = this.snapshot[index].payload.doc.id;
-    this.postDoc = this.db.doc('posts/' + id);
-    this.postsCollection.doc(id).get().subscribe(res => {
-      let post: Post = { ...res.data() } as Post;
-      post.liked = !post.liked;
-      this.postDoc.set(post);
-    }
-    );
-  }
-
-  updateNote(index, note: Post) {
-    console.log('inside updateNote');
-    let id = this.snapshot[index].payload.doc.id;
-    this.postDoc = this.db.doc('posts/' + id);
-    this.postsCollection.doc(id).get().subscribe(res => {
-      this.postDoc.set(note);
-    });
-  }
-}
