--- src/app/pages/personality/personality.component.css
@@ -1,48 +1,0 @@
-/* Style the tab */
-.tab2 {
-    overflow: auto;
-    border: 1px solid #ccc;
-    background-color: #363636;
-    width: 100%;
-  }
-  
-  /* Style the buttons that are used to open the tab content */
-  .tab2 button {
-    background-color: inherit;
-    float: left;
-    border: none;
-    outline: none;
-    cursor: pointer;
-    padding: 14px 14px;
-    transition: 0.3s;
-    color: white;
-  }
-  
-  /* Change background color of buttons on hover */
-  .tab2 button:hover {
-    background-color: #ddd;
-  }
-  
-  /* Create an active/current tablink class */
-  .tab2 button.active {
-    background-color: #ccc;
-  }
-  
-  /* Style the tab content */
-  .tabcontent2 {
-    display: none;
-    padding: 6px 12px;
-    border: 1px solid #ccc;
-    border-top: none;
-  }
-
-  #bottomcon{
-    color: white;
-    width: 100%;
-    background-color: #363636;
-    border: 2px 2px 2px 2px #da4453;
-    border-style: solid;
-    text-align: center;
-    overflow: auto;
-    max-height: 1000px;
-}
