--- src/app/pages/personality/personality.component.html
@@ -1,158 +1,0 @@
-<div id="page" style="margin-left: 2%; margin-right: 2%;">
-  <mat-card>
-    <h1 style="text-align: center;">Profile</h1>
-    <!-- Header info -->
-    <mat-card-header *ngIf="personReady | async" style="border: 1px solid black">
-      <!-- Avatar Image -->
-      <img mat-card-avatar src="{{user.image}}">
-      <mat-card-title>Name: {{user.fname}} {{user.lname}}
-        <button class="btn btn-success" data-toggle="modal" data-target="#subModal" id="subbtn">Subscribe | +</button>
-        <button style="display: none;" class="btn btn-danger" data-toggle="modal" data-target="#deleteModal" id="unsubbtn">Unsubscribe</button>
-      </mat-card-title>
-      <mat-card-subtitle>Email: {{user.email}}</mat-card-subtitle>
-    </mat-card-header>
-    <!-- Image -->
-    <img style="margin-left:0%; width: 100%;" mat-card-image src="https://png.pngtree.com/thumb_back/fw800/back_pic/00/14/61/7856639324bb6f6.jpg" alt="profile image">
-    <!-- Bio and Content -->
-    <mat-card-content id="content-container">
-      <!-- Container -->
-      <div class="tab2">
-        <button id="featuredbtn" class="tablinks2" (click)="tabby2(event, 'FeaturedTab')">Featured</button>
-        <button id="worksbtn" class="tablinks2" (click)="tabby2(event, 'WorksTab')">Filmography/Music</button>
-        <button id="newsbtn" class="tablinks2" (click)="tabby2(event, 'NewsTab')">News</button>
-        <button id="aboutbtn" class="tablinks2" (click)="tabby2(event, 'AboutTab')">About</button>
-        <button id="calendarbtn" class="tablinks2" (click)="tabby2(event, 'CalendarTab')">Calendar and Events</button>
-        <button class="tablinks2" (click)="tabby2(event, '')">X</button>
-      </div>
-      <div class="animated bounceIn" id="bottomcon">
-        <!-- Featured -->
-        <!-- <div id="FeaturedTab" class="tabcontent2" style="width:100%; overflow:auto;">
-          <div style="text-align: center;">Featured</div>
-          <table class="animated bounceIn" id="featured">
-            <tr>
-              <th>Name</th>
-              <td *ngFor="let feature of featuredList">{{feature.name}}</td>
-            </tr>
-            <tr>
-              <th>Image</th>
-              <td *ngFor="let feature of featuredList">
-                <img src="{{feature.image}}" alt="image failed to load" style="height:200px;">
-              </td>
-            </tr>
-          </table>
-        </div> -->
-        <!-- Works -->
-        <!-- <div id="WorksTab" class="tabcontent2" style="width: 100%; overflow: auto;">
-          <div style="text-align: center;">Shop</div>
-          <table class="animated bounceIn" id="shop">
-            <tr>
-              <th>Name</th>
-              <td *ngFor="let product of shopList">{{product.name}}</td>
-            </tr>
-            <tr>
-              <th>Image</th>
-              <td *ngFor="let product of shopList">
-                <img src="{{product.image}}" alt="image failed to load" style="height:200px;">
-              </td>
-            </tr>
-            <tr>
-              <th>Price</th>
-              <td *ngFor="let product of shopList">{{product.price}} <button class="btn btn-success sm" id="buy"
-                  (click)="purchase(id)">Purchase</button>
-              </td>
-            </tr>
-          </table>
-        </div> -->
-        <!-- News -->
-        <div id="NewsTab" class="tabcontent2" style="margin-left:20%; margin-right: 20%; width: 60%; overflow: auto;">
-          <mat-sidenav-container class="example-container">
-            <mat-card class="example-card" *ngFor="let article of mArticles">
-              <mat-card-header>
-                <div mat-card-avatar [ngStyle]="{'background-image': 'url(../assets/images/'+ article.source.id +'.png)'}"
-                  class="example-header-image"></div>
-                <mat-card-title class="title">{{article.title}}</mat-card-title>
-                <mat-card-subtitle>{{article.source.name}}</mat-card-subtitle>
-              </mat-card-header>
-              <img mat-card-image class="img-article" src={{article.urlToImage}} alt="">
-              <mat-card-content>
-                <p>
-                  {{article.description}}
-                </p>
-              </mat-card-content>
-              <mat-card-actions class="action-buttons">
-                <button mat-button color="primary">
-                  <mat-icon>thumb_up_alt</mat-icon> 12 Likes
-                </button>
-                <button mat-button color="primary">
-                  <mat-icon>comment</mat-icon> Comments
-                </button>
-                <button mat-button color="primary">
-                  <mat-icon>share</mat-icon> Share
-                </button>
-                <a mat-button color="primary" href={{article.url}} target="_blank">
-                  <mat-icon>visibility</mat-icon> More
-                </a>
-              </mat-card-actions>
-            </mat-card>
-          </mat-sidenav-container>
-        </div>
-        <!-- About -->
-        <div  *ngIf="personReady | async" id="AboutTab" class="tabcontent2" style="width: 100%; overflow: auto;">
-          <!-- Placeholder using About us page. Not intended to stay. -->
-          {{user.bio}}
-        </div>
-        <!-- Calendar -->
-        <div id="CalendarTab" class="tabcontent2" style="width: 100%; overflow: auto;">
-          <calendar></calendar>
-        </div>
-      </div>
-    </mat-card-content>
-    <!-- Actions? -->
-    <!-- <mat-card-actions>
-      <button mat-button>LIKE</button>
-      <button mat-button>SHARE</button>
-    </mat-card-actions> -->
-  </mat-card>
-</div>
-
-<!-- Subscribe Modal -->
-<div class="modal fade" id="subModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
-  <div class="modal-dialog" role="document">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h5 class="modal-title" id="exampleModalLabel">Confirmation</h5>
-        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-          <span aria-hidden="true">&times;</span>
-        </button>
-      </div>
-      <div class="modal-body">
-        Confirm Subscription? You will be billed for this subscription.
-      </div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-outline-secondary" data-dismiss="modal">Cancel</button>
-        <button type="button" class="btn btn-outline-success" (click)="subscribe()" data-dismiss="modal">Confirm</button>
-      </div>
-    </div>
-  </div>
-</div>
-
-<!-- Unsubscribe Modal -->
-<div class="modal fade" id="deleteModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
-  <div class="modal-dialog" role="document">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h5 class="modal-title" id="exampleModalLabel">Confirmation</h5>
-        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-          <span aria-hidden="true">&times;</span>
-        </button>
-      </div>
-      <div class="modal-body">
-        Unsubscribe? You will not be refunded for this subscription.
-      </div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-outline-secondary" data-dismiss="modal">Cancel</button>
-        <button type="button" class="btn btn-outline-danger" (click)="unsubscribe()" data-dismiss="modal">Confirm</button>
-      </div>
-    </div>
-  </div>
-</div>
