--- src/app/pages/personality/personality.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PersonalityComponent } from './personality.component';
-
-describe('PersonalityComponent', () => {
-  let component: PersonalityComponent;
-  let fixture: ComponentFixture<PersonalityComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ PersonalityComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(PersonalityComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
