+++ src/app/pages/personality2/content/content.component.scss
