+++ src/app/pages/personality2/content/post-dialog/post-dialog.component.html
