+++ src/app/pages/personality2/content/post.ts
